check A

+++ 0x0A-configuration_management/2-execute_a_command.pp
@@ -1,5 +1,6 @@
-# execute a command
-exec { 'pkill':
+# Manifest to kill a process named killmenow using Puppet
+exec { 'killmenow':
   command => 'pkill -f killmenow',
-  path    => '/usr/bin'
+  path    => '/bin:/usr/bin:/sbin:/usr/sbin',
+  onlyif  => 'pgrep -f killmenow',
 }
